requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
asgiref==3.8.1
attrs==24.2.0
alabaster==1.0.0
backoff==2.2.1
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
build==1.2.2
bullet==2.2.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.7
click==8.1.7
cliff==4.7.0
cmd2==2.4.3
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
compressed_rtf==1.0.6
configobj==5.0.8
configparser==7.1.0
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.0.0
docutils==0.21.2
email_validator==2.1.1
easygui==0.98.3
easyocr==1.7.1
ebcdic==1.1.1
exception==0.1.0
extract-msg==0.49.0
faiss-cpu==1.8.0.post1
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gnupg==2.3.1
google-auth==2.34.0
googleapis-common-protos==1.65.0
gotrue==2.4.4
greenlet==3.1.0
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imagesize==1.4.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipwhois==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain==0.3.0
langchain-aws==0.2.2
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.10
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.134
lark==1.1.9
lazy_loader==0.4
logging==0.4.9.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==5.0.0
monotonic==1.6
mpmath==1.3.0
msgspec==0.18.6
msoffcrypto-tool==5.4.2
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
onnxruntime==1.19.2
openai==1.46.0
opencv-python-headless==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pcodedmp==1.2.6
pillow==10.4.0
platformdirs==4.3.2
pluggy==1.5.0
postgrest==0.16.8
posthog==3.6.6
prettytable==3.11.0
protobuf==4.25.5
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.10
pyparsing==3.1.4
pypdf==5.0.0
PyPDF2==3.0.1
PyPika==0.48.9
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
red-black-tree-mod==1.20
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
SQLAlchemy==2.0.35
torch==2.4.1
torchvision==0.19.1
uvicorn==0.30.6
watchfiles==0.24.0
Werkzeug==3.0.4
WTForms==3.1.2
zipp==3.20.2

-e .